% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(degrees)
}
\arguments{
\item{degrees}{# An angle value in units of degrees}
}
\value{
# The value in radian units
}
\description{
Convert degrees to radians
}
\examples{

}
