% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_distance.R
\name{great_circle_distance}
\alias{great_circle_distance}
\title{Calculates distance between two points}
\usage{
great_circle_distance(lat1, lon1, lat2, lon2, scale = 6371)
}
\arguments{
\item{scale}{}
}
\value{
# Distance between the points
}
\description{
Calculates distance between two points
}
